+++ processes/sfrCluster.py
@@ -27,6 +27,7 @@ class ClusterProcess(CoreProcess):
 
         # ElasticSearch Connection
         self.createElasticConnection()
+        self.createElasticSearchIngestPipeline()
         self.createElasticSearchIndex()
 
     def runProcess(self):